front/requirements.txt
discord.py
aioredis==2.0.0a1